Improve code readability by renaming command package Refactor command package name to lowercase
