+++ src/main/java/command/Command.java
@@ -1,4 +1,4 @@
-package Command;
+package command;
 import duke.DukeException;
 import duke.TaskList;
 import duke.UI;
