+++ src/main/java/command/DeleteCommand.java
@@ -1,25 +1,23 @@
-package Command;
+package command;
 
 import duke.DukeException;
 import duke.TaskList;
 import duke.UI;
 import duke.Task;
 
-public class MarkCommand extends Command {
+public class DeleteCommand extends Command {
     private int taskIndex;
 
-    public MarkCommand(int taskIndex) {
+    public DeleteCommand(int taskIndex) {
         this.taskIndex = taskIndex;
     }
 
     @Override
     public String execute(TaskList tasks, UI ui) throws DukeException {
         if (taskIndex < 0 || taskIndex >= tasks.size()) {
-            throw new DukeException("Invalid task number. Index out of bounds.");
+            throw new DukeException("Task number out of bounds.");
         }
-
-        Task task = tasks.get(taskIndex);
-        task.markAsDone();
-        return ui.showTaskMarkedAsDone(task);
+        Task removedTask = tasks.removeTask(taskIndex);
+        return ui.showTaskRemoved(removedTask, tasks.size());
     }
 }
